Create Latest.py
This is an update to the original file. The flat-earth calculation had about a 10% error in the distance calculation. The Haversine calculation has a 0.5% error because it's Haversine and believes the world is a sphere. Vincenty's formula is accurate to within 0.5mm except there's something wrong somewhere because the results are...

Haversine distance in miles  211.26665655709147
distance in miles vincenty  1156.6438391693057
Spherical distance  1157.0134987551473
Distance in Miles using flat earth calculation  222.50855854053702

Of this lot, clearly the Haversine is the distance calculation to use. Vincenty shouldjust have converted from meters to miles by dividing the result by 1,600 but something else seems to be going on. This is a work in progress